[UPGRADE] [DAMAGE] Upgrade system completed and introduced damage to player
The upgrade choosen gets applied to the player and checks for further levels ups until there is not enough exp to level up or all the upgrades have be obtained.
Upgrades are separated into families, families are linear group of upgrades each of which has a level assigned and are unlocked one by one.
Player receives damage when enemies touch him and also repulse to give breathing room to the user.